Translate to Spanish and improves the code's quality of the server and the client
Translates Spanish variables and methods name to English and removes
redundancy of variables and methods with the interface 'Conexión'.

This commit, too, deletes some inconsistency of return types in similar
methods between the Client and Server class, making the functions that
send a specific message, don't return anything. The fix is missing in
the implementations, inside the game interface classes.