Makes inheritance for the graphic interface and the game management
Creates a class that contains all methods in common with the
server, client and singleplayer game, and it creates a game management
class for the methods and variables of the classes that arbitrate the
game.

Also, this commit deletes the class Person, merging it with Player,
to create a global Player. And for the "AI", is now required to pass
a player as an argument in order to take decisions and actions