Fix a bug of #dca1034
That commit made the deck always enabled, and so, the singleplayer
believed that the game had never started when it could be already
running. This commits makes it that, instead of seeing if the deck is
enabled or not, now it's based on whether or not the deck has any
ActionListeners on it.